fix(br): rename backup-manager to tidb-backup-manager
