+++ Makefile
@@ -29,7 +29,7 @@ KIND_VERSION ?= v0.24.0
 # TODO: use kubectl in _output
 KUBECTL = kubectl -n tidb-admin --context kind-tidb-operator
 
-ALL_CMD = tidb-operator prestop-checker testing-workload backup-manager
+ALL_CMD = tidb-operator prestop-checker testing-workload tidb-backup-manager
 .PHONY: build
 build: $(addprefix build/,$(ALL_CMD))
 build/%:
