+++ cmd/tidb-backup-manager/app/backup/backup.go
@@ -30,7 +30,7 @@ import (
 
 	"github.com/pingcap/tidb-operator/api/v2/br/v1alpha1"
 	corev1alpha1 "github.com/pingcap/tidb-operator/api/v2/core/v1alpha1"
-	backupUtil "github.com/pingcap/tidb-operator/cmd/backup-manager/app/util"
+	backupUtil "github.com/pingcap/tidb-operator/cmd/tidb-backup-manager/app/util"
 )
 
 // Options contains the input arguments to the backup command
