+++ cmd/tidb-backup-manager/app/cmd/clean.go
@@ -27,8 +27,8 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/pingcap/tidb-operator/api/v2/br/v1alpha1"
-	"github.com/pingcap/tidb-operator/cmd/backup-manager/app/clean"
-	"github.com/pingcap/tidb-operator/cmd/backup-manager/app/util"
+	"github.com/pingcap/tidb-operator/cmd/tidb-backup-manager/app/clean"
+	"github.com/pingcap/tidb-operator/cmd/tidb-backup-manager/app/util"
 	backupMgr "github.com/pingcap/tidb-operator/pkg/controllers/br/manager/backup"
 	"github.com/pingcap/tidb-operator/pkg/scheme"
 )
