+++ cmd/tidb-backup-manager/app/cmd/restore.go
@@ -29,8 +29,8 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/pingcap/tidb-operator/api/v2/br/v1alpha1"
-	"github.com/pingcap/tidb-operator/cmd/backup-manager/app/restore"
-	"github.com/pingcap/tidb-operator/cmd/backup-manager/app/util"
+	"github.com/pingcap/tidb-operator/cmd/tidb-backup-manager/app/restore"
+	"github.com/pingcap/tidb-operator/cmd/tidb-backup-manager/app/util"
 	restoreMgr "github.com/pingcap/tidb-operator/pkg/controllers/br/manager/restore"
 	"github.com/pingcap/tidb-operator/pkg/scheme"
 )
