+++ cmd/tidb-backup-manager/app/restore/manager.go
@@ -29,22 +29,22 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/pingcap/tidb-operator/api/v2/br/v1alpha1"
-	"github.com/pingcap/tidb-operator/cmd/backup-manager/app/util"
-	"github.com/pingcap/tidb-operator/pkg/controllers/br/manager/restore"
+	"github.com/pingcap/tidb-operator/cmd/tidb-backup-manager/app/util"
 	restoreMgr "github.com/pingcap/tidb-operator/pkg/controllers/br/manager/restore"
 )
 
 type Manager struct {
 	cli           client.Client
-	StatusUpdater restore.RestoreConditionUpdaterInterface
+	StatusUpdater restoreMgr.RestoreConditionUpdaterInterface
 	Options
 }
 
 // NewManager return a RestoreManager
 func NewManager(
 	cli client.Client,
-	statusUpdater restore.RestoreConditionUpdaterInterface,
-	restoreOpts Options) *Manager {
+	statusUpdater restoreMgr.RestoreConditionUpdaterInterface,
+	restoreOpts Options,
+) *Manager {
 	return &Manager{
 		cli,
 		statusUpdater,
