+++ cmd/tidb-backup-manager/app/restore/restore.go
@@ -35,7 +35,7 @@ import (
 
 	"github.com/pingcap/tidb-operator/api/v2/br/v1alpha1"
 	corev1alpha1 "github.com/pingcap/tidb-operator/api/v2/core/v1alpha1"
-	backupUtil "github.com/pingcap/tidb-operator/cmd/backup-manager/app/util"
+	backupUtil "github.com/pingcap/tidb-operator/cmd/tidb-backup-manager/app/util"
 	"github.com/pingcap/tidb-operator/pkg/controllers/br/manager/constants"
 	"github.com/pingcap/tidb-operator/pkg/controllers/br/manager/restore"
 	restoreMgr "github.com/pingcap/tidb-operator/pkg/controllers/br/manager/restore"
