+++ cmd/tidb-backup-manager/main.go
@@ -19,7 +19,7 @@ import (
 
 	"k8s.io/klog/v2"
 
-	"github.com/pingcap/tidb-operator/cmd/backup-manager/app"
+	"github.com/pingcap/tidb-operator/cmd/tidb-backup-manager/app"
 )
 
 func main() {
